Moved some functionality from LearningMode to AppStateManager
This completes the refactoring of AppState.cs from before.
